+++ src/cljs/cxengage_javascript_sdk/modules/entities.cljs
@@ -31,10 +31,9 @@
   [params]
   (let [{:keys [callback topic resource-id]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/users/resource-id")
         get-user-request {:method :get
-                          :url (iu/build-api-url-with-params
-                                url
+                          :url (iu/api-url
+                                "tenants/tenant-id/users/resource-id"
                                 {:tenant-id tenant-id
                                  :resource-id resource-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-user-request))]
@@ -57,10 +56,9 @@
   [params]
   (let [{:keys [callback topic]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/users")
         get-users-request {:method :get
-                           :url (iu/build-api-url-with-params
-                                 url
+                           :url (iu/api-url
+                                 "tenants/tenant-id/users"
                                  {:tenant-id tenant-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-users-request))]
     (when (= status 200)
@@ -84,10 +82,9 @@
   [params]
   (let [{:keys [callback topic queue-id]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/queues/queue-id")
         get-queue-request {:method :get
-                           :url (iu/build-api-url-with-params
-                                 url
+                           :url (iu/api-url
+                                 "tenants/tenant-id/queues/queue-id"
                                  {:tenant-id tenant-id
                                   :queue-id queue-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-queue-request))]
@@ -110,10 +107,9 @@
   [params]
   (let [{:keys [callback topic]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/queues")
         get-queues-request {:method :get
-                           :url (iu/build-api-url-with-params
-                                 url
+                           :url (iu/api-url
+                                 "tenants/tenant-id/queues"
                                  {:tenant-id tenant-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-queues-request))]
     (when (= status 200)
@@ -137,10 +133,9 @@
   [params]
   (let [{:keys [callback topic transfer-list-id]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/transfer-lists/transfer-list-id")
         get-transfer-list-request {:method :get
-                                   :url (iu/build-api-url-with-params
-                                         url
+                                   :url (iu/api-url
+                                         "tenants/tenant-id/transfer-lists/transfer-list-id"
                                          {:tenant-id tenant-id
                                           :transfer-list-id transfer-list-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-transfer-list-request))]
@@ -163,10 +158,9 @@
   [params]
   (let [{:keys [callback topic]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/transfer-lists")
         get-transfer-lists-request {:method :get
-                                    :url (iu/build-api-url-with-params
-                                          url
+                                    :url (iu/api-url
+                                          "tenants/tenant-id/transfer-lists"
                                           {:tenant-id tenant-id})}
         {:keys [status api-response]} (a/<! (iu/api-request get-transfer-lists-request))]
     (when (= status 200)
@@ -195,11 +189,10 @@
   [params]
   (let [{:keys [callback topic update-body resource-id]} params
         tenant-id (st/get-active-tenant-id)
-        url (str (st/get-base-api-url) "tenants/tenant-id/users/resource-id")
         put-user-request {:method :put
                           :body update-body
-                          :url (iu/build-api-url-with-params
-                                url
+                          :url (iu/api-url
+                                "tenants/tenant-id/users/resource-id"
                                 {:tenant-id tenant-id
                                  :resource-id resource-id})}
         {:keys [status api-response]} (a/<! (iu/api-request put-user-request))]